ci-artifacts: sort minimal-sdk's test jobs in reverse runtime order
When running 17 jobs in parallel, it makes sense to start the
longest-running one first, then the second-longest-running etc. That
way, the overall runtime is minimized.

This commit orders the jobs by their _current_ runtimes, in descending
order, as measured in a single run with the current `master` of git.git.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>